requirements.txt
pyyaml==5.3.1
tensorflow==2.8.0
spektral==1.0.8